token-verifier/requirements.txt
boto3==1.38.10
PyJWT==2.10.1
